[FIX] account_payment_term_cutoff_day: fix not changing to end of month
Before this commit, automatically shifting the due date by one month did not account
for the end of the month. For example, if the original due date was the 30th and the
following month had 31 days, the new due date would incorrectly be set to the 30th
instead of the 31st.